Update README.md
-Scaled down d_model from to 1024 -> 3.059
-Changed from Muon to NorMuon -> 3.0396.
-Finally got extra value embeddings mixed with the V-matrix to work. Got identical results using learnable scalar mixing of embeddings in only the final two layers as when mixing in all layers, so ended up going with that implementation. -> 3.032
-Increased the amount of training steps to the edge of the time-limit. -> 3.0305. 
I changed to measuring validation loss against a small subset of the validation set during training. I calculate validation loss on entire validation set after training instead, which is not included in the runtime. The run shown in the graph actually reached 3.0285 in validation loss, but to have some margin for worse GPU performance I have decreased my number of steps by 200 for my submission. This gives a validation loss to 3.0305.